requirements.txt
pillow
cryptography
customtkinter
numpy
matplotlib
tkinterdnd2